Update fib.py
 Test the function with the following value. My name is Karen , so the first and last letter of my name (K+ N = 11 + 14) give the number 25. The 25th Fibonacci number is 75025.
Fibonacci number 25 is 75025
I had some issues. I followed the helpful thread so thanks to all contributors. 


My surname is Byrne

The first letter B is number 66

The last letter e is number 101

Fibonacci number 167 is 35600075545958458963222876581316753

Ord: Given a string of length one, return an integer representing the Unicode code point of the character when the argument is a unicode object, or the value of the byte when the argument is an 8-bit string.